backend-python/requirement.txt
Flask==2.3.3
PyPDF2==3.0.1
google-generativeai==0.3.2
# Let Flask decide the right Werkzeug version
Werkzeug>=2.3.7
flask_cors
dotenv